move the verkle commands to their own file
